Allow leading/trailing slashes
This allows routes in form ////item/////// to still match.
This can be useful as sometimes people make errors in parsing
url strings.  While they should be corrected, it's not an illegal
thing to perform when requesting links.

Allow the ability of chaining setters so we don't need to run:
$route->map('GET', '/someroute/', 'somefunction', 'name');
$route->map('PUT', '/someotherroute/', 'someotherfunction', 'someothername');

We can just do:
$route
    ->map('GET', '/someroute/', 'somefunction', 'name')
    ->map('PUT', '/someotherroute/', 'someotherfunciton', 'someothername');